requirements.txt
streamlit==1.45.0
langchain==0.3.25
langchain-google-genai==2.0.0
langchain-huggingface==0.2.0
langchain_community
faiss-cpu==1.10.0
sentence-transformers==4.1.0
python-dotenv==1.1.0
tiktoken==0.9.0
PyMuPDF==1.25.5
google-generativeai==0.7.2